refactor: extract year and segment parsing
